+++ Lib/TSCompiler/ProjectOptions.cs
@@ -50,6 +50,7 @@ namespace Lib.TSCompiler
         public bool NoHtml;
         public bool WarningsAsErrors;
         public bool PreserveProjectRoot;
+        public bool ForceNewBundler;
         public string JasmineVersion;
         public List<string> TestDirectories;
         public string PathToTranslations;
@@ -505,6 +506,7 @@ namespace Lib.TSCompiler
                 ProcessEnvs!["NODE_ENV"] = Parser.Parse("DEBUG?\"development\":\"production\"");
             }
             PreserveProjectRoot = bobrilSection?["preserveProjectRoot"]?.Value<bool>() ?? false;
+            ForceNewBundler = bobrilSection?["forceNewBundler"]?.Value<bool>() ?? false;
         }
 
         Dictionary<string, AstToplevel> ParseDefines(JObject? jObject)